+++ src/Symfony/Component/Validator/Tests/ValidationTest.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Validator\Tests;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Validator\Constraints\Blank;
+use Symfony\Component\Validator\Constraints\Email;
 use Symfony\Component\Validator\Constraints\NotBlank;
 use Symfony\Component\Validator\Exception\ValidationFailedException;
 use Symfony\Component\Validator\Validation;
@@ -41,6 +42,39 @@ class ValidationTest extends TestCase
             $this->assertCount(1, $violations);
             $this->assertEquals('This value should be blank.', $violations->get(0)->getMessage());
         }
+    }
+
+    public function testInvokableValid()
+    {
+        $validation = new Validation(new NotBlank(), new Email());
+        $this->assertEquals('test@example.com', $validation('test@example.com'));
+    }
+
+    public function testInvokableInvalid()
+    {
+        $validation = new Validation(new Blank());
+        try {
+            $validation('test');
+            $this->fail('No ValidationFailedException thrown');
+        } catch (ValidationFailedException $e) {
+            $this->assertEquals('test', $e->getValue());
+
+            $violations = $e->getViolations();
+            $this->assertCount(1, $violations);
+            $this->assertEquals('This value should be blank.', $violations->get(0)->getMessage());
+        }
+    }
+
+    public function testInvokableWithValidator()
+    {
+        $validation = new Validation(Validation::createValidator(), new NotBlank());
+        $this->assertEquals('test@example.com', $validation('test@example.com'));
+    }
+
+    public function testInvokableWithoutConstraints()
+    {
+        $validation = new Validation();
+        $this->assertEquals('test', $validation('test'));
     }
 
     public function testCreateIsValidCallableValid()